Extract ItemSelector component
Extract a generic ItemSelector from AddDropdownButton. ItemSelector holds the dropdown/filter/sort logic but no longer hardcodes the anchor: callers pass a render-prop child that receives the menu handler and returns their own anchor element, and the tooltip label is now a required prop. AddDropdownButton becomes a thin wrapper that supplies its button anchor and tooltip.

Prepares for #10688.

Closes #13653.
